Refactored code and fixed bug in displaying medium-long bodies
